requirements.txt
openai>=1.109.1,<3.0.0
langchain
langchain-openai
langchain-community
faker
pandas
sqlalchemy
mysql-connector-python
pymysql
python-dotenv
qdrant-client
Flask
PyPDF2
